Populate path filters env variables using the CLI
