+++ lib/analyze.js
@@ -33,7 +33,6 @@ const perf_hooks_1 = require("perf_hooks");
 const toolrunner = __importStar(require("@actions/exec/lib/toolrunner"));
 const del_1 = __importDefault(require("del"));
 const yaml = __importStar(require("js-yaml"));
-const analysisPaths = __importStar(require("./analysis-paths"));
 const codeql_1 = require("./codeql");
 const configUtils = __importStar(require("./config-utils"));
 const feature_flags_1 = require("./feature-flags");
@@ -79,9 +78,6 @@ async function setupPythonExtractor(logger, features, codeql) {
     process.env["LGTM_PYTHON_SETUP_VERSION"] = output;
 }
 async function createdDBForScannedLanguages(codeql, config, logger, features) {
-    // Insert the LGTM_INDEX_X env vars at this point so they are set when
-    // we extract any scanned languages.
-    analysisPaths.includeAndExcludeAnalysisPaths(config);
     for (const language of config.languages) {
         if ((0, languages_1.isScannedLanguage)(language) &&
             !dbIsFinalized(config, language, logger)) {
