+++ lib/init.js
@@ -28,7 +28,6 @@ const fs = __importStar(require("fs"));
 const path = __importStar(require("path"));
 const toolrunner = __importStar(require("@actions/exec/lib/toolrunner"));
 const safeWhich = __importStar(require("@chrisgavin/safe-which"));
-const analysisPaths = __importStar(require("./analysis-paths"));
 const codeql_1 = require("./codeql");
 const configUtils = __importStar(require("./config-utils"));
 const languages_1 = require("./languages");
@@ -45,7 +44,7 @@ exports.initCodeQL = initCodeQL;
 async function initConfig(languagesInput, queriesInput, packsInput, configFile, dbLocation, configInput, trapCachingEnabled, debugMode, debugArtifactName, debugDatabaseName, repository, tempDir, codeQL, workspacePath, gitHubVersion, apiDetails, logger) {
     logger.startGroup("Load language configuration");
     const config = await configUtils.initConfig(languagesInput, queriesInput, packsInput, configFile, dbLocation, configInput, trapCachingEnabled, debugMode, debugArtifactName, debugDatabaseName, repository, tempDir, codeQL, workspacePath, gitHubVersion, apiDetails, logger);
-    analysisPaths.printPathFiltersWarning(config, logger);
+    printPathFiltersWarning(config, logger);
     logger.endGroup();
     return config;
 }
@@ -67,6 +66,14 @@ async function runInit(codeql, config, sourceRoot, processName, registriesInput,
     return await (0, tracer_config_1.getCombinedTracerConfig)(config);
 }
 exports.runInit = runInit;
+function printPathFiltersWarning(config, logger) {
+    // Index include/exclude/filters only work in javascript/python/ruby.
+    // If any other languages are detected/configured then show a warning.
+    if ((config.paths.length !== 0 || config.pathsIgnore.length !== 0) &&
+        !config.languages.every(languages_1.isScannedLanguage)) {
+        logger.warning('The "paths"/"paths-ignore" fields of the config only have effect for JavaScript, Python, and Ruby');
+    }
+}
 /**
  * Possibly convert this error into a UserError in order to avoid
  * counting this error towards our internal error budget.
